+++ src/queries/index.ts
@@ -1,2 +1,3 @@
 export * from './useUserInfoQuery';
 export * from './useQuizListInfinityQuery';
+export * from './useLogoutQuery';
